use 32bits instead of 8bits for the om and string number encoding in tableio
git-svn-id: http://code.icecube.wisc.edu/svn/projects/tableio/trunk@84985 16731396-06f5-0310-8873-f7f720988828

+++ private/tableio/converter/I3WaveformConverter.cxx
@@ -48,8 +48,8 @@ I3TableRowDescriptionPtr I3WaveformConverter::CreateDescription(const I3Waveform
     }
     
     desc->isMultiRow_ = true;
-    desc->AddField<int8_t>("string", "", "String number");
-    desc->AddField<uint8_t>("om", "", "OM number");
+    desc->AddField<int32_t>("string", "", "String number");
+    desc->AddField<uint32_t>("om", "", "OM number");
     
     desc->AddField<bool>("atwd_ok", "bool", "status flag that atwd waveform has been convertered");
     desc->AddField<double>("atwd_t0", "ns", "start time of ATWD waveform");
